+++ lib/extensions.dart
@@ -4,5 +4,7 @@ export './extensions/list.dart';
 export './extensions/navigation.dart';
 export './extensions/snackbar.dart';
 export './extensions/strings.dart';
-export 'extensions/text_styles.dart';
+export './extensions/text_styles.dart';
 export './extensions/theme.dart';
+export './extensions/widget.dart';
+export './extensions/number.dart';